Enable logging using xng-rs-log
Add flag to force linking against XRE event if linking against LithOS.
This is required so that XalPrintf is available in the partition.

Add an attribute to pass patches to the LithOS-Ops package. This is
required so that the linker script can be patched to allow for extra
BSS sections. Otherwise, some static variables are not zero-initialized.
An example for this is a static variable that holds XalLogger from
xng-rs-log.